requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.1.3
gcloud==0.18.3
googleapis-common-protos==1.56.4
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jws==0.1.3
MarkupSafe==2.1.1
numpy==1.23.1
oauth2client==4.1.3
pandas==1.4.3
protobuf==4.21.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.15.0
pyparsing==3.0.9
Pyrebase4==4.5.0
python-dateutil==2.8.2
python-jwt==2.0.1
pytz==2022.1
requests==2.28.1
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
urllib3==1.26.10
Werkzeug==2.1.2
zipp==3.8.1
